use explicit cast to avoid windows warning

+++ libraries/workload/src/workload/Space.cpp
@@ -67,8 +67,8 @@ void Space::setViews(const std::vector<Space::View>& views) {
 }
 
 void Space::categorizeAndGetChanges(std::vector<Space::Change>& changes) {
-    uint32_t numProxies = _proxies.size();
-    uint32_t numViews = _views.size();
+    uint32_t numProxies = (uint32_t)_proxies.size();
+    uint32_t numViews = (uint32_t)_views.size();
     for (uint32_t i = 0; i < numProxies; ++i) {
         Proxy& proxy = _proxies[i];
         if (proxy.region < Space::REGION_INVALID) {
